Exploded types_cloud into separate files
