+++ ovh/types_cloud_project_kube.go
@@ -1,0 +1,70 @@
+package ovh
+
+import (
+	"fmt"
+	"strings"
+
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/ovh/terraform-provider-ovh/ovh/helpers"
+)
+
+type CloudProjectKubeCreateOpts struct {
+	Name             *string `json:"name,omitempty"`
+	PrivateNetworkId *string `json:"privateNetworkId,omitempty"`
+	Region           string  `json:"region"`
+	Version          *string `json:"version,omitempty"`
+}
+
+func (opts *CloudProjectKubeCreateOpts) FromResource(d *schema.ResourceData) *CloudProjectKubeCreateOpts {
+	opts.Region = d.Get("region").(string)
+	opts.Version = helpers.GetNilStringPointerFromData(d, "version")
+	opts.Name = helpers.GetNilStringPointerFromData(d, "name")
+	opts.PrivateNetworkId = helpers.GetNilStringPointerFromData(d, "private_network_id")
+
+	return opts
+}
+
+func (s *CloudProjectKubeCreateOpts) String() string {
+	return fmt.Sprintf("%s(%s): %s", *s.Name, s.Region, *s.Version)
+}
+
+type CloudProjectKubeResponse struct {
+	ControlPlaneIsUpToDate bool     `json:"controlPlaneIsUpToDate"`
+	Id                     string   `json:"id"`
+	IsUpToDate             bool     `json:"isUpToDate"`
+	Name                   string   `json:"name"`
+	NextUpgradeVersions    []string `json:"nextUpgradeVersions"`
+	NodesUrl               string   `json:"nodesUrl"`
+	PrivateNetworkId       string   `json:"privateNetworkId"`
+	Region                 string   `json:"region"`
+	Status                 string   `json:"status"`
+	UpdatePolicy           string   `json:"updatePolicy"`
+	Url                    string   `json:"url"`
+	Version                string   `json:"version"`
+}
+
+func (v CloudProjectKubeResponse) ToMap() map[string]interface{} {
+	obj := make(map[string]interface{})
+	obj["control_plane_is_up_to_date"] = v.ControlPlaneIsUpToDate
+	obj["id"] = v.Id
+	obj["is_up_to_date"] = v.IsUpToDate
+	obj["name"] = v.Name
+	obj["next_upgrade_versions"] = v.NextUpgradeVersions
+	obj["nodes_url"] = v.NodesUrl
+	obj["private_network_id"] = v.PrivateNetworkId
+	obj["region"] = v.Region
+	obj["status"] = v.Status
+	obj["update_policy"] = v.UpdatePolicy
+	obj["url"] = v.Url
+	obj["version"] = v.Version[:strings.LastIndex(v.Version, ".")]
+
+	return obj
+}
+
+func (s *CloudProjectKubeResponse) String() string {
+	return fmt.Sprintf("%s(%s): %s", s.Name, s.Id, s.Status)
+}
+
+type CloudProjectKubeKubeConfigResponse struct {
+	Content string `json:"content"`
+}
