+++ ovh/types_cloud_project_networkprivate.go
@@ -1,0 +1,72 @@
+package ovh
+
+import (
+	"fmt"
+)
+
+// Opts
+type CloudProjectNetworkPrivateCreateOpts struct {
+	ServiceName string   `json:"serviceName"`
+	VlanId      int      `json:"vlanId"`
+	Name        string   `json:"name"`
+	Regions     []string `json:"regions"`
+}
+
+func (p *CloudProjectNetworkPrivateCreateOpts) String() string {
+	return fmt.Sprintf("projectId: %s, vlanId:%d, name: %s, regions: %s", p.ServiceName, p.VlanId, p.Name, p.Regions)
+}
+
+// Opts
+type CloudProjectNetworkPrivateUpdateOpts struct {
+	Name string `json:"name"`
+}
+
+type CloudProjectNetworkPrivateRegion struct {
+	Status string `json:"status"`
+	Region string `json:"region"`
+}
+
+func (p *CloudProjectNetworkPrivateRegion) String() string {
+	return fmt.Sprintf("Status:%s, Region: %s", p.Status, p.Region)
+}
+
+type CloudProjectNetworkPrivateResponse struct {
+	Id      string                              `json:"id"`
+	Status  string                              `json:"status"`
+	Vlanid  int                                 `json:"vlanId"`
+	Name    string                              `json:"name"`
+	Type    string                              `json:"type"`
+	Regions []*CloudProjectNetworkPrivateRegion `json:"regions"`
+}
+
+func (p *CloudProjectNetworkPrivateResponse) String() string {
+	return fmt.Sprintf("Id: %s, Status: %s, Name: %s, Vlanid: %d, Type: %s, Regions: %s", p.Id, p.Status, p.Name, p.Vlanid, p.Type, p.Regions)
+}
+
+// Opts
+type CloudProjectNetworkPrivatesCreateOpts struct {
+	ServiceName string `json:"serviceName"`
+	NetworkId   string `json:"networkId"`
+	Dhcp        bool   `json:"dhcp"`
+	NoGateway   bool   `json:"noGateway"`
+	Start       string `json:"start"`
+	End         string `json:"end"`
+	Network     string `json:"network"`
+	Region      string `json:"region"`
+}
+
+func (p *CloudProjectNetworkPrivatesCreateOpts) String() string {
+	return fmt.Sprintf("PCPNSCreateOpts[projectId: %s, networkId:%s, dhcp: %v, noGateway: %v, network: %s, start: %s, end: %s, region: %s]",
+		p.ServiceName, p.NetworkId, p.Dhcp, p.NoGateway, p.Network, p.Start, p.End, p.Region)
+}
+
+type CloudProjectNetworkPrivatesResponse struct {
+	Id        string    `json:"id"`
+	GatewayIp string    `json:"gatewayIp"`
+	Cidr      string    `json:"cidr"`
+	IPPools   []*IPPool `json:"ipPools"`
+}
+
+func (p *CloudProjectNetworkPrivatesResponse) String() string {
+	return fmt.Sprintf("PCPNSResponse[Id: %s, GatewayIp: %s, Cidr: %s, IPPools: %s]", p.Id, p.GatewayIp, p.Cidr, p.IPPools)
+}
